Change all floating point arithmetic from double to float.
Some microcontrollers have single precision floating point hardware.


git-svn-id: svn+ssh://subversion.munts.net/src/trunk/ARM@24517 06ee2f96-5407-11de-8bab-f3d7589a4122

+++ arduino/RP2040/SparkFunProMicro/test_pwm/test_pwm.ino
@@ -39,7 +39,7 @@ void setup()
 
 void loop()
 {
-  for (double duty = DUTYCYCLE_MIN; duty <= DUTYCYCLE_MAX; duty += 0.5)
+  for (float duty = DUTYCYCLE_MIN; duty <= DUTYCYCLE_MAX; duty += 0.5F)
   {
     outp = duty;
     delay(50);
